fix: move country zoom hook to ZoomControl (previously it did run before the map was actually defined)
